requirements.txt
browser-use==0.5.11
langchain_openai
python-dotenv
playwright